Split supervisor configurations
Laravel configurations split up so we can mix and match by copying to `/etc/supervisor/conf.d`. No need to comment anything out here.
We commented out queue worker redis usage as I personally would take the approach of using Horizon with Redis but the database for other queue work. Typically the queue workers I've been exposed to start out as database and "graduate" to redis or horizon. Some developers are also more comfortable looking in one place for the information instead of two.
There is no `artisan notification:worker` that I know of from Greg Sanderson's examples but my guess is it should be a queue worker dedicated to that queue. The other queue worker example doesn't list defaults like we have it but I find it better to be explicit with values we'll likely change.

For reference, artisan queue:work defaults:
```
      --name[=NAME]          The name of the worker [default: "default"]
      --queue[=QUEUE]        The names of the queues to work
      --daemon               Run the worker in daemon mode (Deprecated)
      --once                 Only process the next job on the queue
      --stop-when-empty      Stop when the queue is empty
      --delay[=DELAY]        The number of seconds to delay failed jobs (Deprecated) [default: "0"]
      --backoff[=BACKOFF]    The number of seconds to wait before retrying a job that encountered an uncaught exception [default: "0"]
      --max-jobs[=MAX-JOBS]  The number of jobs to process before stopping [default: "0"]
      --max-time[=MAX-TIME]  The maximum number of seconds the worker should run [default: "0"]
      --force                Force the worker to run even in maintenance mode
      --memory[=MEMORY]      The memory limit in megabytes [default: "128"]
      --sleep[=SLEEP]        Number of seconds to sleep when no job is available [default: "3"]
      --rest[=REST]          Number of seconds to rest between jobs [default: "0"]
      --timeout[=TIMEOUT]    The number of seconds a child process can run [default: "60"]
      --tries[=TRIES]        Number of times to attempt a job before logging it failed [default: "1"]
```